Replacing src files and moving back public into the matchscouting folder
