+++ MatchScouting/public/manifest.json
